hyundai_document_authenticator/external/result_gui/requirements.txt
Flask>=2.3.0,<3.0.0
Flask-Login>=0.6.2,<0.7.0
bcrypt>=4.0.1,<5.0.0
psycopg2-binary>=2.9.9,<3.0.0
PyYAML>=6.0.1,<7.0.0
Typer>=0.9.0,<1.0.0
pyotp>=2.9.0,<3.0.0
